+++ Modules/bindings.py
@@ -146,8 +146,24 @@ def rebind_Clusters( self, NWKID):
     for iterBindCluster in cluster_to_bind:
         for iterEp in self.ListOfDevices[NWKID]['Ep']:
             if iterBindCluster in self.ListOfDevices[NWKID]['Ep'][iterEp]:
-                loggingBinding( self, 'Debug', 'Request a Bind  for %s/%s on Cluster %s' %(NWKID, iterEp, iterBindCluster), nwkid=NWKID)
+                loggingBinding( self, 'Debug', 'Request a rebind for %s/%s on Cluster %s' %(NWKID, iterEp, iterBindCluster), nwkid=NWKID)
                 bindDevice( self, self.ListOfDevices[NWKID]['IEEE'], iterEp, iterBindCluster)
+
+def reWebBind_Clusters( self, NWKID):
+
+    if 'WebBind' in self.ListOfDevices[NWKID]:
+        for Ep in list(self.ListOfDevices[NWKID]['WebBind']):
+            for cluster in list(self.ListOfDevices[NWKID]['WebBind'][ Ep ]):
+                for destNwkid in list(self.ListOfDevices[NWKID]['WebBind'][ Ep ][cluster]):
+                    if destNwkid in ('Stamp','Target','TargetIEEE','SourceIEEE','TargetEp','Phase','Status'): # delete old mechanism
+                        Domoticz.Error("---> delete  destNwkid: %s" %( destNwkid))
+                        del self.ListOfDevices[NWKID]['WebBind'][Ep][cluster][destNwkid]
+                    if self.ListOfDevices[NWKID]['WebBind'][Ep][cluster][destNwkid]['Phase'] == 'binded':
+                        loggingBinding( self, 'Debug', "Request a rewebbind for : nwkid %s ep: %s cluster: %s destNwkid: %s" 
+                            %(NWKID,Ep,cluster,destNwkid), nwkid=NWKID)
+                        self.ListOfDevices[NWKID]['WebBind'][Ep][cluster][destNwkid]['Stamp'] = int(time())
+                        self.ListOfDevices[NWKID]['WebBind'][Ep][cluster][destNwkid]['Phase'] = 'requested'
+                        return
 
 def unbindDevice( self, ieee, ep, cluster, destaddr=None, destep="01"):
     '''
@@ -301,5 +317,33 @@ def isWebBind( self, sourceIeee, sourceEp, destIeee, destEp, Cluster):
                         return True
     return False
 
+def callBackForWebBindIfNeeded( self , srcNWKID ):
 
+    """
+    Check that WebBind are well set
+    """
 
+    if srcNWKID not in self.ListOfDevices:
+        return
+    if 'WebBind' not in self.ListOfDevices[srcNWKID]:
+        return
+
+    for Ep in list(self.ListOfDevices[srcNWKID]['WebBind']):
+        for ClusterId in list(self.ListOfDevices[srcNWKID]['WebBind'][ Ep ]):
+            for destNwkid in list(self.ListOfDevices[srcNWKID]['WebBind'][ Ep ][ClusterId]):
+                if destNwkid in ('Stamp','Target','TargetIEEE','SourceIEEE','TargetEp','Phase','Status'):
+                    Domoticz.Error("---> delete  destNwkid: %s" %( destNwkid))
+                    del self.ListOfDevices[srcNWKID]['WebBind'][Ep][ClusterId][destNwkid]
+                elif ('Phase' in self.ListOfDevices[srcNWKID]['WebBind'][Ep][ClusterId][destNwkid] and \
+                                 self.ListOfDevices[srcNWKID]['WebBind'][Ep][ClusterId][destNwkid]['Phase'] == 'requested'):
+                    if ('Stamp' in self.ListOfDevices[srcNWKID]['WebBind'][Ep][ClusterId][destNwkid] and \
+                        time() < self.ListOfDevices[srcNWKID]['WebBind'][Ep][ClusterId][destNwkid]['Stamp']+ 5):    # Let's wait 5s before trying again
+                        continue
+                    loggingBinding( self, 'Log', "Redo a WebBind for device %s" %(srcNWKID))
+                    sourceIeee = self.ListOfDevices[srcNWKID]['WebBind'][Ep][ClusterId][destNwkid]['SourceIEEE']
+                    destIeee = self.ListOfDevices[srcNWKID]['WebBind'][Ep][ClusterId][destNwkid]['TargetIEEE']
+                    destEp = self.ListOfDevices[srcNWKID]['WebBind'][Ep][ClusterId][destNwkid]['TargetEp']
+                    # Perforning the bind
+                    webBind(self, sourceIeee, Ep, destIeee, destEp, ClusterId)
+
+                    self.ListOfDevices[srcNWKID]['WebBind'][Ep][ClusterId][destNwkid]['Stamp'] = int(time())
